App refactor: Introduce AnimationContextProvider
Improve App module's read- and compose-ability by moving AnimationContext
concerns to a separate provider module.
